+++ config/webpack.package.js
@@ -5,6 +5,7 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const commonConfig = require('./webpack.common');
 const { ENV, dir, APP_VERSION } = require('./helpers');
 // const ngtools = require('@ngtools/webpack');
+const { CheckerPlugin } = require('awesome-typescript-loader');
 
 const banner =
 `/**
@@ -22,18 +23,10 @@ module.exports = function(env) {
         {
           test: /\.ts$/,
           loaders: [
-            'awesome-typescript-loader'
+            'awesome-typescript-loader',
+            'angular2-template-loader'
           ],
           exclude: [/\.(spec|e2e|d)\.ts$/]
-        },
-        {
-          test: /\.scss$/,
-          loaders: [
-            'style-loader',
-            'css-loader?sourceMap',
-            'postcss-loader?sourceMap',
-            'sass-loader?sourceMap'
-          ]
         }
       ]
     },
@@ -65,17 +58,17 @@ module.exports = function(env) {
       'd3-selection': 'd3-selection',
       'd3-shape': 'd3-shape',
       'd3-hierarchy': 'd3-hierarchy',
+      'd3-time': 'd3-time',
+      'd3-time-format': 'd3-time-format',
       'rxjs': 'rxjs',
       'rxjs/Rx': 'rxjs/Rx',
-      'rxjs/Subject': 'rxjs/Subject',
-      'rxjs/Subscription': 'rxjs/Subscription',
-      'rxjs/observable/PromiseObservable': 'rxjs/observable/PromiseObservable',
-      'rxjs/operator/toPromise': 'rxjs/operator/toPromise',
       'rxjs/Observable': 'rxjs/Observable',
-      'zone.js/dist/zone': 'zone.js/dist/zone',
-      'moment': 'moment'
+      'rxjs/add/observable/fromEvent': 'rxjs/add/observable/fromEvent',
+      'rxjs/add/operator/debounceTime': 'rxjs/add/operator/debounceTime',
+      'zone.js/dist/zone': 'zone.js/dist/zone'
     },
     plugins: [
+      new CheckerPlugin(),
       new webpack.BannerPlugin({
         banner: banner,
         raw: true,