+++ release/heat-map/heat-map.component.d.ts
@@ -1,3 +1,4 @@
+
 import { BaseChartComponent } from '../common/base-chart.component';
 import { ViewDimensions } from '../common/view-dimensions.helper';
 import { ColorHelper } from '../common/color.helper';
@@ -6,10 +7,15 @@ export declare class HeatMapComponent extends BaseChartComponent {
     xAxis: any;
     yAxis: any;
     showXAxisLabel: any;
+    yDomainMin: any;
+    yDomainMax: any;
     showYAxisLabel: any;
     xAxisLabel: any;
     yAxisLabel: any;
     gradient: boolean;
+    innerPadding: Number | Number[];
+    xAxisTickFormatting: any;
+    yAxisTickFormatting: any;
     dims: ViewDimensions;
     xDomain: any[];
     yDomain: any[];