+++ src/common/legend/legend.component.ts
@@ -1,13 +1,6 @@
 import {
-  Component,
-  Input,
-  ChangeDetectionStrategy,
-  Output,
-  EventEmitter,
-  SimpleChanges,
-  OnChanges,
-  ChangeDetectorRef,
-  NgZone
+  Component, Input, ChangeDetectionStrategy, Output, EventEmitter,
+  SimpleChanges, OnChanges, ChangeDetectorRef, NgZone, ViewEncapsulation
  } from '@angular/core';
  import { formatLabel } from '../label.helper';
 
@@ -39,6 +32,8 @@ import {
       </div>
     </div>
   `,
+  styleUrls: ['./legend.component.scss'],
+  encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class LegendComponent implements OnChanges {
@@ -70,12 +65,12 @@ export class LegendComponent implements OnChanges {
   }
 
   getLegendEntries(): any[] {
-    let items = [];
+    const items = [];
 
     for(const label of this.data) {
       const formattedLabel = formatLabel(label);
 
-      let idx = items.findIndex((i) => {
+      const idx = items.findIndex((i) => {
         return i.label === formattedLabel;
       });
 
@@ -93,7 +88,7 @@ export class LegendComponent implements OnChanges {
 
   isActive(entry): boolean {
     if(!this.activeEntries) return false;
-    let item = this.activeEntries.find(d => {      
+    const item = this.activeEntries.find(d => {      
       return entry.label === d.name;
     });
     return item !== undefined;