+++ src/gauge/gauge.component.ts
@@ -6,7 +6,8 @@ import {
   AfterViewInit,
   ChangeDetectionStrategy,
   Output,
-  EventEmitter
+  EventEmitter,
+  ViewEncapsulation
 } from '@angular/core';
 
 import d3 from '../d3';
@@ -48,7 +49,8 @@ import { ColorHelper } from '../common/color.helper';
           [radius]="outerRadius"
           [angleSpan]="angleSpan"
           [valueScale]="valueScale"
-          [startAngle]="startAngle">
+          [startAngle]="startAngle"
+          [tickFormatting]="axisTickFormatting">
         </svg:g>
 
         <svg:text #textEl
@@ -62,6 +64,11 @@ import { ColorHelper } from '../common/color.helper';
       </svg:g>
     </ngx-charts-chart>
   `,
+  styleUrls: [
+    '../common/base-chart.component.scss',
+    './gauge.component.scss'
+  ],
+  encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class GaugeComponent extends BaseChartComponent implements AfterViewInit {
@@ -69,6 +76,7 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
   @Input() legend = false;
   @Input() min: number = 0;
   @Input() max: number = 100;
+  @Input() textValue: string;
   @Input() units: string;
   @Input() bigSegments: number = 10;
   @Input() smallSegments: number = 5;
@@ -77,6 +85,7 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
   @Input() startAngle: number = -120;
   @Input() angleSpan: number = 240;
   @Input() activeEntries: any[] = [];
+  @Input() axisTickFormatting: any;
 
   // Specify margins
   @Input() margin: any[];
@@ -149,8 +158,8 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
       this.setColors();
       this.legendOptions = this.getLegendOptions();
 
-      let xOffset = this.margin[3] + this.dims.width / 2;
-      let yOffset = this.margin[0] + this.dims.height / 2;
+      const xOffset = this.margin[3] + this.dims.width / 2;
+      const yOffset = this.margin[0] + this.dims.height / 2;
 
       this.transform = `translate(${ xOffset }, ${ yOffset })`;
       this.rotation = `rotate(${ this.startAngle })`;
@@ -159,21 +168,21 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
   }
 
   getArcs(): any[] {
-    let arcs = [];
+    const arcs = [];
 
-    let availableRadius = this.outerRadius * 0.7;
+    const availableRadius = this.outerRadius * 0.7;
 
-    let radiusPerArc = Math.min(availableRadius / this.results.length, 10);
-    let arcWidth = radiusPerArc * 0.7;
+    const radiusPerArc = Math.min(availableRadius / this.results.length, 10);
+    const arcWidth = radiusPerArc * 0.7;
     this.textRadius = this.outerRadius - this.results.length * radiusPerArc;
     this.cornerRadius = Math.floor(arcWidth / 2);
 
     let i = 0;
-    for (let d of this.results) {
-      let outerRadius = this.outerRadius - (i * radiusPerArc);
-      let innerRadius = outerRadius - arcWidth;
+    for (const d of this.results) {
+      const outerRadius = this.outerRadius - (i * radiusPerArc);
+      const innerRadius = outerRadius - arcWidth;
 
-      let backgroundArc = {
+      const backgroundArc = {
         endAngle: this.angleSpan * Math.PI / 180,
         innerRadius,
         outerRadius,
@@ -183,7 +192,7 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
         }
       };
 
-      let valueArc = {
+      const valueArc = {
         endAngle: Math.min(this.valueScale(d.value), this.angleSpan) * Math.PI / 180,
         innerRadius,
         outerRadius,
@@ -193,7 +202,7 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
         }
       };
 
-      let arc = {
+      const arc = {
         backgroundArc,
         valueArc
       };
@@ -210,9 +219,9 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
   }
 
   getValueDomain(): any[] {
-    let values = this.results.map(d => d.value);
-    let dataMin = Math.min(...values);
-    let dataMax = Math.max(...values);
+    const values = this.results.map(d => d.value);
+    const dataMin = Math.min(...values);
+    const dataMax = Math.max(...values);
 
     if (this.min !== undefined) {
       this.min = Math.min(this.min, dataMin);
@@ -232,11 +241,16 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
   getValueScale(): any {
     return d3.scaleLinear()
       .range([0, this.angleSpan])
+      .nice()
       .domain(this.valueDomain);
   }
 
   getDisplayValue(): string {
-    let value = this.results.map(d => d.value).reduce((a, b) => { return a + b; }, 0);
+    const value = this.results.map(d => d.value).reduce((a, b) => a + b, 0);
+
+    if(this.textValue && 0 !== this.textValue.length) {
+      return this.textValue.toLocaleString();
+    }
     return value.toLocaleString();
   }
 
@@ -247,7 +261,7 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
 
       if (width === 0) {
         this.resizeScale = 1;
-      } else {        
+      } else {
         const availableSpace = this.textRadius;
         this.resizeScale = Math.floor((availableSpace / (width / this.resizeScale)) * 100) / 100;
       }
@@ -303,7 +317,7 @@ export class GaugeComponent extends BaseChartComponent implements AfterViewInit 
 
   isActive(entry): boolean {
     if(!this.activeEntries) return false;
-    let item = this.activeEntries.find(d => {
+    const item = this.activeEntries.find(d => {
       return entry.name === d.name && entry.series === d.series;
     });
     return item !== undefined;