+++ src/pie-chart/pie-chart.component.ts
@@ -2,6 +2,7 @@ import {
   Component,
   Input,
   Output,
+  ViewEncapsulation,
   EventEmitter,
   ChangeDetectionStrategy
 } from '@angular/core';
@@ -37,6 +38,11 @@ import { BaseChartComponent } from '../common/base-chart.component';
       </svg:g>
     </ngx-charts-chart>
   `,
+  styleUrls: [
+    '../common/base-chart.component.scss',
+    './pie-chart.component.scss'
+  ],
+  encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class PieChartComponent extends BaseChartComponent {
@@ -45,6 +51,7 @@ export class PieChartComponent extends BaseChartComponent {
   @Input() legend = false;
   @Input() explodeSlices = false;
   @Input() doughnut = false;
+  @Input() arcWidth = 0.25;
   @Input() gradient: boolean;
   @Input() activeEntries: any[] = [];
 
@@ -78,8 +85,8 @@ export class PieChartComponent extends BaseChartComponent {
         columns: 10
       });
 
-      let xOffset = this.margin[3] + this.dims.width / 2;
-      let yOffset = this.margin[0] + this.dims.height / 2;
+      const xOffset = this.margin[3] + this.dims.width / 2;
+      const yOffset = this.margin[0] + this.dims.height / 2;
       this.translation = `translate(${xOffset}, ${yOffset})`;
       this.outerRadius = Math.min(this.dims.width, this.dims.height);
       if (this.labels) {
@@ -90,7 +97,7 @@ export class PieChartComponent extends BaseChartComponent {
       }
       this.innerRadius = 0;
       if (this.doughnut) {
-        this.innerRadius = this.outerRadius * 0.75;
+        this.innerRadius = this.outerRadius * (1 - this.arcWidth);
       }
 
       this.domain = this.getDomain();
@@ -106,7 +113,7 @@ export class PieChartComponent extends BaseChartComponent {
   }
 
   getDomain(): any[] {
-    let items = [];
+    const items = [];
 
     this.results.map(d => {
       let label = d.name;